pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vozctl"
version = "0.1.0"
description = "Voice control for developers"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{ name = "LFG Consulting" }]

dependencies = [
    "sherpa-onnx>=1.10",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "pynput>=1.7",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
]

[project.scripts]
vozctl = "vozctl.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vozctl"]
